+++ mooncake-transfer-engine/src/transport/CMakeLists.txt
@@ -1,9 +1,12 @@
 file(GLOB XPORT_SOURCES "*.cpp")
 
 add_subdirectory(rdma_transport)
+add_subdirectory(rdma_twosided)
 add_subdirectory(rpc_communicator)
-add_library(transport OBJECT ${XPORT_SOURCES} $<TARGET_OBJECTS:rdma_transport>
-                             $<TARGET_OBJECTS:rpc_communicator>)
+add_library(
+  transport OBJECT
+  ${XPORT_SOURCES} $<TARGET_OBJECTS:rdma_transport>
+  $<TARGET_OBJECTS:rdma_twosided> $<TARGET_OBJECTS:rpc_communicator>)
 target_link_libraries(
   transport PRIVATE JsonCpp::JsonCpp yalantinglibs::yalantinglibs glog::glog
                     pthread)
@@ -92,7 +95,9 @@ if(USE_CXI)
   target_link_libraries(transport PRIVATE fabric)
 endif()
 
-if(USE_CUDA OR USE_MUSA OR USE_MACA)
+if(USE_CUDA
+   OR USE_MUSA
+   OR USE_MACA)
   add_subdirectory(device)
   target_sources(transport PUBLIC $<TARGET_OBJECTS:device_transport>)
   # device_transport (ibgda_device_transport.cpp / mlx5gda.cpp) calls libmlx5
